refactor(LottoController): 이동 및 LottoMaker 삭제
LottoMaker의 기능을 LottoController로 이동하여 LottoController가 단일 책임을 가질 수 있도록 구조를 정리하고 LottoMaker 클래스를 제거했습니다.
